We're going to need to rip it out.
All the LML, Autumn in particular, needs to get out of here. I am sick of it. There is no documentation for even very important things. The classpath scanning has been the source of serious bugs and I need to stay using an older version because all of the related libraries are so buggy. AGH.